fixed broken migration

+++ db/migrate/20110215014234_bootstrap.rb
@@ -26,7 +26,7 @@ class Bootstrap < ActiveRecord::Migration
       t.integer    :need,       :null => false, :default => 0
       t.integer    :greed,      :null => false, :default => 0
       t.integer    :loyalty,    :null => false, :default => 0
-      t.float      :score,      :null => false, :default => 0
+      t.decimal    :score,      :null => false, :default => 0
 
       t.timestamps
     end
